legacy/programm/requirements.txt
PyQt5==5.15.10
numpy>=1.26.0
opencv-python>=4.9.0
Pillow>=9.0.0
requests>=2.25.0
PyYAML>=6.0
onnxruntime>=1.16.0
torch>=2.0.0
torchvision>=0.15.0
